EmbedPneumoXRay/Devtool_SSHChain/requirements.txt
pyside6 # GPL 3.0
pyyaml # MIT
paramiko # LGPL v2.1
ansi2html # GPL v3.0

# Development
black # MIT
